Add script to get time since last Bitcoin block
I found myself constantly wanting to see how long it had been since a
Bitcoin block was mined. It became a mess of a one-liner so I decided to
let it have it's own script.

The script must be run from a full node with a running installation of
`bitcoind` so it can query the information it needs.