Add script to disable dock badges on Mac
I find it difficult to maintain enough self-control to ignore the little
dock notification badges when I'm supposed to be working. Even when I go
into Do Not Disturb the badges are still displayed. This little shell
script edits some settings to turn off those badges.

Note that it only works on macOS.