Add diffchars tool
I hacked together a quick Python script to find the number of different
characters between two strings. It looks at each character index of each
string in turn and then adds one to a counter if the characters don't
match. I use it for finding the number of differences between DNA
strands for biology.