Add licence Python script
Added a script to print out a specified licence. You can then do anthing
you like with it, like use regex to replace the placeholders with the
current year and your name, or pipe it into a file.

I start new coding projects all the time and got tired of constantly
looking up the licence that I wanted to use on the web and copy pasting
that into the licence file so I made this script that stores the most
popular licences for easy access.